LayoutTests/fast/images/slower-animation-than-decoding-image-expected.txt

Ensure the image frame duration is respected even if the frame finishes decoding early.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.imageFrameIndex(image) is 2
PASS successfullyParsed is true

TEST COMPLETE
